Split away Language.Dart.GrammarDeclaration
