+++ language-dart.cabal
@@ -21,6 +21,7 @@ library
   exposed-modules:     Language.Dart.Grammar
                        Language.Dart.Pretty
                        Language.Dart.Syntax
+  other-modules:       Language.Dart.GrammarDeclaration
   build-depends:       base >= 4.7 && < 5
                      , grammatical-parsers >= 0.3 && < 0.4
                      , parsers >= 0.12.7 && < 0.13
